+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.1.2 under development
 
-- no changes in this release.
+- Enh #74: Dispatch `MissingTranslationCategoryEvent` once per category (@vjik)
+- Enh #73: Set `en_US` as default locale for translator (@vjik)
+- Enh #72: Use `SimpleMessageFormatter` to format messages when missing translation category (@vjik)
 
 ## 1.1.1 September 09, 2022